tmf: Fix Double state value type
Reviewed-on: https://git.eclipse.org/r/16346
IP-Clean: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Hudson CI

+++ org.eclipse.linuxtools.tmf.core/src/org/eclipse/linuxtools/tmf/core/statevalue/DoubleStateValue.java
@@ -27,7 +27,7 @@ final class DoubleStateValue extends TmfStateValue {
 
     @Override
     public Type getType() {
-        return Type.INTEGER;
+        return Type.DOUBLE;
     }
 
     @Override
